customerservice entities updated

+++ customerservice/src/main/java/com/etiya/customerservice/dto/address/CreateAddressRequestDto.java
@@ -24,5 +24,5 @@ public class CreateAddressRequestDto {
     @NotEmpty
     private String street;
     @NotNull
-    private Integer houseNumber;
+    private Long houseNumber;
 }
